fix: moving accounts to service folder
